normalize act

+++ dev/compute_mse_l1_validation_loss.py
@@ -1,4 +1,3 @@
-import contextlib
 import math
 import os
 import shutil
@@ -44,9 +43,10 @@ def evaluate_checkpoint(policy, validation_loader, device):
             batch["observation.image"] = torch.squeeze(batch["observation.image"], dim=1)
             batch["action"] = torch.squeeze(batch["action"], dim=1)
         else:
-            batch["action"] = torch.squeeze(batch["action"])
+            batch["action"] = torch.squeeze(batch["action"], dim=1)
         observation = {key: batch[key].to(device, non_blocking=True) for key in keys}
         batch["action"] = batch["action"].to(device, non_blocking=True)
+        policy.to(device)
         batch = policy.normalize_targets(batch)
         actions = batch["action"]
         with torch.no_grad():
@@ -58,8 +58,8 @@ def evaluate_checkpoint(policy, validation_loader, device):
 
 
 def main(hydra_cfg, list_of_dirs):
-    model_json = "/fsx/marina_barannikov/outputs/train/compare_val_loss/pusht_84/020000/config.json"
-    device = hydra_cfg.device
+    model_json = "/fsx/marina_barannikov/outputs/train/compare_val_loss/transfer_cube_84/010000/config.json"
+    device = "cuda"
 
     val_dataset = get_train_val_datasets(hydra_cfg, split_value=0.8)
     validation_loader = torch.utils.data.DataLoader(val_dataset, batch_size=1, shuffle=False)
@@ -71,7 +71,7 @@ def main(hydra_cfg, list_of_dirs):
                 subdir_path = os.path.join(root_dir, subdir)
                 checkpoint_path = os.path.join(subdir_path, "model.safetensors")
                 if os.path.isfile(checkpoint_path):
-                    with contextlib.suppress(shutil.SameFileError):
+                    if not os.path.exists(os.path.join(subdir_path, "config.json")):
                         shutil.copy(model_json, subdir_path)
                     print(f"Evaluating checkpoint from {checkpoint_path}")
                     policy = make_policy(hydra_cfg=hydra_cfg, pretrained_policy_name_or_path=subdir_path)
@@ -83,7 +83,7 @@ def main(hydra_cfg, list_of_dirs):
     df = pd.DataFrame(results)
 
     # Save DataFrame to a CSV file
-    output_file = "dev/pusht_normalized_results.csv"
+    output_file = "dev/transfer_cube_normalized_results.csv"
     df.to_csv(output_file, index=False)
 
     print(f"Results saved to {output_file}")
@@ -93,14 +93,16 @@ def main(hydra_cfg, list_of_dirs):
 
 if __name__ == "__main__":
     list_of_dirs = [
-        "/fsx/marina_barannikov/outputs/train/compare_val_loss/pusht_85",
-        "/fsx/marina_barannikov/outputs/train/compare_val_loss/pusht_84",
-        "/fsx/marina_barannikov/outputs/train/compare_val_loss/pusht_100000",
+        "/fsx/marina_barannikov/outputs/train/compare_val_loss/transfer_cube_84",
+        "/fsx/marina_barannikov/outputs/train/compare_val_loss/transfer_cube_85",
+        "/fsx/marina_barannikov/outputs/train/compare_val_loss/transfer_cube_1000",
     ]
-    policy_path = Path("/fsx/marina_barannikov/outputs/train/compare_val_loss/pusht_84/010000")
+    policy_path = Path("/fsx/marina_barannikov/outputs/train/compare_val_loss/transfer_cube_84/010000")
     cfg = init_hydra_config(str(policy_path / "config.yaml"))
     if cfg.policy.name == "diffusion":
         cfg.training.delta_timestamps["observation.state"] = [0]
         cfg.training.delta_timestamps["observation.image"] = [0]
         cfg.training.delta_timestamps["action"] = [i / 10 for i in range(1)]
+    else:
+        cfg.training.delta_timestamps["action"] = [0]
     main(cfg, list_of_dirs)
